[Refactor] 고객 요리사 리스트 페이지 리팩토링
